Button: Move related files into folder, refactor
