+++ FrontEnd/src/components/Button/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './Button';
+export * from './Button';
